Change API rate limiting.
Instead of using a single rate limit bucket for searching by uploading
an image and sending a hash, use two separate buckets joined together.

Now when an image is uploaded, it consumes an image and a hash.
When just a hash is provided, it only consumes an image. This naming is
somewhat confusing, but was used for data backwards compatibility.